+++ packages/core/src/Models/Cart.php
@@ -2,6 +2,7 @@
 
 namespace Lunar\Models;
 
+use Exception;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Casts\AsArrayObject;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
@@ -40,6 +41,7 @@ use Lunar\Validation\Cart\ValidateCartForOrderCreation;
 /**
  * @property int $id
  * @property ?int $user_id
+ * @property ?int $customer_id
  * @property ?int $merged_id
  * @property int $currency_id
  * @property int $channel_id
@@ -208,6 +210,16 @@ class Cart extends BaseModel
     public function user()
     {
         return $this->belongsTo(config('auth.providers.users.model'));
+    }
+
+    /**
+     * Return the customer relationship.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
+     */
+    public function customer()
+    {
+        return $this->belongsTo(Customer::class);
     }
 
     public function scopeUnmerged($query)
@@ -452,10 +464,36 @@ class Cart extends BaseModel
      */
     public function associate(User $user, $policy = 'merge', $refresh = true)
     {
+        if ($this->customer()->exists()) {
+            if (! $user->query()
+                ->whereHas('customers', fn ($query) => $query->where('customer_id', $this->customer->id))
+                ->exists()) {
+                throw new Exception('Invalid user');
+            }
+        }
+
         return app(
             config('lunar.cart.actions.associate_user', AssociateUser::class)
         )->execute($this, $user, $policy)
             ->then(fn () => $refresh ? $this->refresh()->calculate() : $this);
+    }
+
+    /**
+     * Associate a customer to the cart
+     */
+    public function setCustomer(Customer $customer): Cart
+    {
+        if ($this->user()->exists()) {
+            if (! $customer->query()
+                ->whereHas('users', fn ($query) => $query->where('user_id', $this->user->id))
+                ->exists()) {
+                throw new Exception('Invalid customer');
+            }
+        }
+
+        $this->customer()->associate($customer)->save();
+
+        return $this->refresh()->calculate();
     }
 
     /**