+++ packages/core/tests/Unit/Managers/StorefrontSessionManagerTest.php
@@ -3,12 +3,16 @@
 namespace Lunar\Tests\Unit\Managers;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Session;
 use Lunar\Base\StorefrontSessionInterface;
+use Lunar\Exceptions\CustomerNotBelongsToUserException;
 use Lunar\Managers\StorefrontSessionManager;
 use Lunar\Models\Channel;
 use Lunar\Models\Currency;
+use Lunar\Models\Customer;
 use Lunar\Models\CustomerGroup;
+use Lunar\Tests\Stubs\User as StubUser;
 use Lunar\Tests\TestCase;
 
 /**
@@ -17,6 +21,11 @@ use Lunar\Tests\TestCase;
 class StorefrontSessionManagerTest extends TestCase
 {
     use RefreshDatabase;
+
+    private function setAuthUserConfig()
+    {
+        Config::set('auth.providers.users.model', 'Lunar\Tests\Stubs\User');
+    }
 
     /**
      * @test
@@ -81,6 +90,36 @@ class StorefrontSessionManagerTest extends TestCase
         $manager = app(StorefrontSessionInterface::class);
 
         $this->assertEquals($currency->id, $manager->getCurrency()->id);
+    }
+
+    /**
+     * @test
+     */
+    public function can_initialise_the_customer()
+    {
+        Channel::factory()->create([
+            'default' => true,
+        ]);
+
+        $this->setAuthUserConfig();
+
+        $user = StubUser::factory()->create();
+
+        $customers = Customer::factory(5)->create();
+
+        $user->customers()->sync($customers->pluck('id'));
+
+        $this->assertCount(5, $user->customers()->get());
+
+        $this->assertDatabaseCount((new Customer)->getTable(), 5);
+
+        $manager = app(StorefrontSessionInterface::class);
+
+        $this->assertNull($manager->getCustomer());
+
+        $this->actingAs($user);
+
+        $this->assertEquals($customers->last()->id, $manager->getCustomer()->id);
     }
 
     /**
@@ -158,5 +197,59 @@ class StorefrontSessionManagerTest extends TestCase
         );
 
         $this->assertCount(1, $manager->getCustomerGroups());
+    }
+
+    /**
+     * @test
+     */
+    public function can_set_customer()
+    {
+        Channel::factory()->create([
+            'default' => true,
+        ]);
+
+        $this->setAuthUserConfig();
+
+        $user = StubUser::factory()->create();
+
+        // $this->actingAs($user);
+
+        $customers = Customer::factory(5)->create();
+
+        $user->customers()->sync($customers->pluck('id'));
+
+        $manager = app(StorefrontSessionInterface::class);
+
+        $customer = $customers->first();
+
+        $manager->setCustomer($customer);
+
+        $this->assertEquals($customer->id, $manager->getCustomer()->id);
+    }
+
+    /**
+     * @test
+     */
+    public function ensure_customer_belongs_to_user()
+    {
+        Channel::factory()->create([
+            'default' => true,
+        ]);
+
+        $this->setAuthUserConfig();
+
+        $user = StubUser::factory()->create();
+
+        $this->actingAs($user);
+
+        $customers = Customer::factory(5)->create();
+
+        $manager = app(StorefrontSessionInterface::class);
+
+        $customer = $customers->first();
+
+        $this->expectException(CustomerNotBelongsToUserException::class);
+
+        $manager->setCustomer($customer);
     }
 }